requirements.txt
pandas
numpy
matplotlib
opencv-python
pillow
torch
torchvision
transformers
lpips
scikit-image
openai
boto3
diffusers
scipy
requests
